requirements/base.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-oauth-toolkit==1.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
oauth2-provider==0.0
oauthlib==3.1.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
